successfully use the chosen path
