+++ index.html
@@ -40,6 +40,7 @@
   </div>
 </div>
 
-<script src="src/renderer.js"></script>
+<script src="src/ngma.js" type="text/javascript"></script>
+<script src="src/renderer.js" type="text/javascript"></script>
 </body>
 </html>
