+++ package.json
@@ -32,6 +32,7 @@
     "glob": "^7.1.2",
     "node-sloc": "^0.1.11",
     "parse-gitignore": "^0.5.1",
+    "path": "^0.12.7",
     "ts-node": "^7.0.0",
     "ts-node-dev": "^1.0.0-pre.26",
     "ts-simple-ast": "^19.1.0",
